@@ -1,7 +1,7 @@
 package com.address.api.service;
 
 import com.address.api.model.mapper.AdressMapper;
-import com.address.api.controller.AdressFeingService;
+import com.address.api.client.AddressClient;
 import com.address.api.model.dto.AdressDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,13 +10,13 @@ import org.springframework.stereotype.Service;
 public class AdressService {
 
     @Autowired
-    private AdressFeingService adressFeingService;
+    private AddressClient addressClient;
 
     @Autowired
     private AdressMapper adressMapper;
 
     public AdressDTO getAdress(String cep){
-        return adressMapper.adressToAdressDTO(adressFeingService.getAdress(cep));
+        return adressMapper.adressToAdressDTO(addressClient.getAddress(cep));
     }
 
 }
